Create Cargo-Bot Interpreter
Cargo-Bot is a programming videogame consisting of vertical stacks of crates and a claw to move the crates from one stack to another. The aim of the game is to move the crates from an initial position to a desired final position.

The player controls the claw by placing commands into one of four programs. The player presses a "play" button when they are ready and the commands are executed. If, after executing the commands, the crates are in the desired final position, the player wins.

Your task in this kata will be to run the player's commands with a given initial state for the crates and return the ending state of the crates.

Crates

Crates are objects having a color field which contains one of 4 possible strings: "RED", "YELLOW", "GREEN", or "BLUE".

The initial state of the board is given as a 2D list of crates, where each inner list represents a stack of crates. For example, the following would represent three stacks with a red crate on top of a yellow crate in the left-most stack and one blue crate in the middle stack:

initial_state = [
  [Crate("YELLOW"), Crate("RED")],
  [Crate("BLUE")],
  []
]
or

  |
| ^     |
|       |
| R     |
| Y B _ |
Programs & Commands

The player controls the claw by placing commands into one of four programs. The programs are provided as a 2D list of commands where each list represents one of the four programs.

Commands are objects with 2 fields: command which represents an action to be performed and flag which represents an optional condition (described in detail below).

The interpreter always starts with the first command in the first program and executes each command in sequence until the end of the program. Programs 2, 3, and 4 are only run if called from the first program.

There are seven different commands: 3 commands for moving the claw and 4 commands for calling programs

Claw Commands

"LEFT" - Moves the claw one position to the left (if the claw is already on the left-most stack, nothing happens)
"RIGHT" - Moves the claw one position to the right (if the claw is already on the right-most stack, nothing happens)
"DOWN" - Does one of three things depending on its state:
If the claw is empty and there are no crates in the stack below, nothing happens.
If the claw is empty and there are crates in the stack below, the claw picks up the top crate.
If the claw is holding a crate, the claw drops the crate on top of the stack below.
For example, the following program would pick up a crate, move the claw to the right, and set down the crate:

program = [Command("DOWN"), Command("RIGHT"), Command("DOWN")]
After 3 steps, this program would produce the following:

  |                       |
| ^     |             |   ^   |
|       |     -->     |       |
| R     |             |   R   |
| Y B _ |             | Y B _ |
Note: the claw always starts above the left-most stack.

Program Commands

In addition to the claw commands, there are four program commands: "PROG1", "PROG2", "PROG3", and "PROG4". Each of these commands directs the interpreter to begin reading commands from the beginning of the corresponding program. Once all the commands from the called program are run, the interpreter returns to the calling program and completes any remaining commands.

For example, the following set of programs would pick up two crates and move them one position to the right:

programs = [
  [Command("PROG2"), Command("LEFT"), Command("PROG2")],
  [Command("DOWN"), Command("RIGHT"), Command("DOWN")],
  [],
  []
]
After 9 steps, these programs would produce the following:

  |                       |
| ^     |             |   ^   |
|       |             |       |
|       |     -->     |   Y   |
| R     |             |   R   |
| Y B _ |             | _ B _ |
Conditional Flags

Each command can also have a conditional flag which instructs the interpreter to only execute the command if the condition is met. The possible flag values together with the conditions they represent are listed below:

"NONE" - the claw is not holding anything.
"ANY" - the claw is holding a crate.
"RED" - the claw is holding a red crate.
"YELLOW" - the claw is holding a yellow crate.
"GREEN" - the claw is holding a green crate.
"BLUE" - the claw is holding a blue crate.
For example, the programs below would move each red crate on the top of each stack one step to the left:

programs = [
  [Command("RIGHT"), Command("PROG2"), Command("PROG1")],
  [Command("DOWN"), Command("PROG3", "RED"), Command("DOWN", "ANY")],
  [Command("LEFT"), Command("DOWN"), Command("RIGHT")],
  []
]
After 42 steps, these programs would produce the following:

  |                                      |
| ^            |             |           ^  |
|              |             |              |
|       R      |     -->     | R            |
| B R   G     R|             | B   R G      |
| Y B B G B _ Y|             | Y B B G B R Y|
Counting Steps

Since many programs will run indefinitely, you will be provided with a maximum number of steps. Your interpreter should return the final state of the board after either

No more commands are left to be run
You have performed the maximum number of steps
Every command counts as a step, even if the command does nothing ("DOWN" with no crate present or "LEFT"/"RIGHT" against a wall) or was not executed due to its conditional flag.